Initial Commit
Created a basic engine for pygame as well as started on creating the
AStar algorithm base.
Includes:

'System' class which currently holds some 'should be constant' variables
to help for debugging later on by putting all configuration variables in
one place. Currently separated into 'Color' variables and 'Display'
variables.

'FPS' class which can control FPS to a certain value and display the
current FPS centered to the window.

'Graphics' which handles setting up a window based on the
'System.Display' configuration variables. Is also used to draw surfaces
to the screen, and any clearing and flipping that needs to be done.
'Graphics.Draw' is the draw function, and will be the most used later on
for reference.

Created a workspace for notepad++ editing. Keeps files together to be
opened much easier in notepad++

Updated gitignore to ignore python compiled files. No thanks