The Heuristic Commit
Indentations fixed to work with Visual Studio. Pretty much means it
can't be edited in notepad++ now...but hey

Nodes now store a rect of their position on the screen so that drawing
the start and goal rectangles is easier. The 'Node' class now has an
'IsTraversable' property so that the 'm_Color' attribute is always
updated to match. More properties will probably be added.

The singleton 'Graphics' class now scales anything drawn through its
functions to the proper value to match a difference in resolution to
window size. It can do this with Surfaces, Rectangles and Lines thus
far. Added a few comments at the top

New 'SearchSpace' class. Handles drawing the searchspace as well as
setting up the searchspace itself. Responds to mouse input. Left click
to make a node in-traversable and right click to move the goal. Holds a
surface for text and the nodes which are only updated when something
changes.

Added the 'Graph' class to 'System' to hold configuration values for the
searchspace's dimensions and node sizes, etc...