The Generator Function Commit
Added more colors to 'System.Color'

Added 'KeyDownListener' in 'Core.py'

Added an outline to the FPS text by drawing multiple black text at an
offset behind the original. Not very efficient in pygame, but is the
only way. Remove for performance.

Completed the basic A* algorhythm in Python.

Added a 'm_DebugSurface' which draws lines to each node's parent and the
ending path. Can be turned on and off, but is not programmed to toggle
yet.

Created a Generator Function for the path finding function. Execution
cannot be speed up or slowed down, but the function currently 'steps'
each frame. Goes around 15 steps per second. Will and variable time
later.

Added an 'UpdateColor' function for the 'Node' class. This ensures that
each node is the correct color. Called right before rendering each
rectangle in 'SearchSpace.Draw'

-----------------------------------------------------------------------
-Generator produce data for iteration
-Coroutines are consumers of data. Coroutines are not related to
iteration