Added Comments for Graphics.py
Added comments for most of the 'Graphics.py' file.

Tested different resolutions, node sizes, node separations, and null
heuristic values. Something with the math on how I'm calculating node
with breaks when inputing a 1 for node size and anything other than 0
for the 'LINE_WIDTH' and 'LINE_HEIGHT'.

Issues also occur when the m_SolveSpeed variable is set to 0. Has
something to do with how the generator function is set up. Need to
figure out what's actually causing the problem.